+++ src/classes/InstructionStack.ts
@@ -15,12 +15,15 @@ export class InstructionStack {
   /**
    * Adds a new instruction to the stack.
    * @author CasperSocio
-   * @version 0.0.3
+   * @version 0.0.5
    * @param instruction The new instruction to add
    * @since 0.0.3
    * @public
    */
-  add(instruction: IInstruction) {
+  add(instruction: IInstruction, value?: number | string) {
+    if (value) {
+      instruction.value = value
+    }
     this._stack.push(instruction)
   }
 
@@ -55,16 +58,28 @@ export class InstructionStack {
   /**
    * Finds matching instructions by name and assigns a new value.
    * @author CasperSocio
-   * @version 0.0.3
+   * @version 0.0.5
    * @param name The instruction target name
    * @param value The new value to assign instruction
+   * @param callback The callback function
    * @since 0.0.3
    * @public
    */
-  findAndReplaceValue(name: TInstructionName, callback: (value: string) => string) {
+  findAndReplaceValue({ name, value, callback }: {
+    name: TInstructionName
+    value?: number | string
+    callback?: (newValue: string) => string
+  }): void {
     this._stack = this._stack.map(instruction => {
       if (instruction.name === name) {
-        instruction.value = callback(instruction.value || '')
+        if (value !== undefined) {
+          instruction.value = value
+        }
+        else if (callback !== undefined) {
+          if (typeof instruction.value === 'string') {
+            instruction.value = callback(instruction.value || '')
+          }
+        }
       }
       return instruction
     })